Moved key lesc and dhkey generation from test_passkey to ble_adapter.
