+++ pc_ble_driver_py/ble_adapter.py
@@ -46,9 +46,56 @@ from pc_ble_driver_py.ble_driver import *
 from pc_ble_driver_py.exceptions import NordicSemiException
 from pc_ble_driver_py.observers import *
 
+from cryptography.hazmat.backends import default_backend
+from cryptography.hazmat.primitives.asymmetric import ec
+
 logger = logging.getLogger(__name__)
 
 MAX_TRIES = 10  # Maximum Number of Tries by driver.ble_gattc_write
+
+
+def change_endianness(input_string):
+    output = ""
+    for i in range(len(input_string) - 1, 0, -2):
+        output += input_string[i - 1]
+        output += input_string[i]
+    return output
+
+
+class PublicKey(object):
+
+    def __init__(self, x, y, curve_type):
+        assert isinstance(x, (int)), 'Invalid argument type'
+        assert isinstance(y, (int)), 'Invalid argument type'
+        self.x = x
+        self.y = y
+        self.curve_type = curve_type
+
+    def to_c(self):
+
+        logger.debug("ECDH key: Before: x: {:X}".format(self.x))
+        logger.debug("ECDH key: Before: y: {:X}".format(self.y))
+
+        x_list = self._int_to_list(self.x)[::-1]
+        y_list = self._int_to_list(self.y)[::-1]
+
+        logger.debug("ECDH key: After: x: {}".format(" ".join(["0x{:02X}".format(i) for i in x_list])))
+        logger.debug("ECDH key: After: y: {}".format(" ".join(["0x{:02X}".format(i) for i in y_list])))
+
+        return util.list_to_uint8_array(x_list + y_list)
+
+    def _int_to_list(self, input_integer):
+        output_list = []
+        input_hex_string = "{:X}".format(input_integer)
+
+        # Add zeros to start key if they are stripped.
+        while len(input_hex_string) < 64:
+            input_hex_string = "0" + input_hex_string
+
+        for i in range(1, len(input_hex_string), 2):
+            output_list.append(int((input_hex_string[i-1] + input_hex_string[i]), 16))
+
+        return output_list
 
 
 class DbConnection(object):
@@ -592,6 +639,73 @@ class BLEAdapter(BLEDriverObserver):
                 else:
                     raise e
         raise NordicSemiException("Unable to successfully call ble_gattc_write")
+
+    def generate_lesc_private_key(self):
+        self.lesc_private_key = ec.generate_private_key(ec.SECP256R1(), default_backend())
+    
+    def generate_lesc_keyset(self, private_key=None):
+        if not private_key:
+            private_key = self.lesc_private_key
+        lesc_own_public_key = private_key.public_key()
+        # Put own lesc public key into keyset.
+        lesc_own_public_key_numbers = lesc_own_public_key.public_numbers()
+        lesc_own_public_key_array = PublicKey(lesc_own_public_key_numbers.x,
+                                              lesc_own_public_key_numbers.y,
+                                              lesc_own_public_key_numbers.curve).to_c()
+
+        self.lesc_pk_own.pk = lesc_own_public_key_array.cast()
+        lesc_own_public_key_list = util.uint8_array_to_list(self.lesc_pk_own.pk, 64)
+        logger.debug("lesc_own_public_key_list {}".format(lesc_own_public_key_list))
+
+        p_pk_own_array = util.list_to_uint8_array(lesc_own_public_key_list)
+        pk = p_pk_own_array.cast()
+
+        keyset = driver.ble_gap_sec_keyset_t()
+
+        keyset.keys_own.p_enc_key = driver.ble_gap_enc_key_t()
+        keyset.keys_own.p_id_key = driver.ble_gap_id_key_t()
+        keyset.keys_own.p_sign_key = driver.ble_gap_sign_info_t()
+        keyset.keys_own.p_pk = driver.ble_gap_lesc_p256_pk_t()
+        keyset.keys_own.p_pk.pk = pk
+
+        keyset.keys_peer.p_enc_key = driver.ble_gap_enc_key_t()
+        keyset.keys_peer.p_id_key = driver.ble_gap_id_key_t()
+        keyset.keys_peer.p_sign_key = driver.ble_gap_sign_info_t()
+        keyset.keys_peer.p_pk = driver.ble_gap_lesc_p256_pk_t()
+        
+        self._keyset = keyset
+        
+        return keyset
+
+    def generate_lesc_dhkey(self, peer_public_key):
+        lesc_pk_peer = peer_public_key
+        # Translate incoming peer public key to x an y components as integers.
+        peer_public_key_list = lesc_pk_peer.pk
+        peer_public_key_x = int(change_endianness("".join(["{:02X}".format(i) for i in peer_public_key_list[:32]])), 16)
+        peer_public_key_y = int(change_endianness("".join(["{:02X}".format(i) for i in peer_public_key_list[32:]])), 16)
+
+        logger.debug("Peer public DH key, big endian, x: 0x{:X}, y: 0x{:X}".format(peer_public_key_x, peer_public_key_y))
+
+        # Generate a _EllipticCurvePublicKey object of the received peer public key.
+        lesc_peer_public_key_obj = ec.EllipticCurvePublicNumbers(peer_public_key_x,
+                                                                 peer_public_key_y,
+                                                                 ec.SECP256R1())
+        lesc_peer_public_key_obj2 = lesc_peer_public_key_obj.public_key(default_backend())
+
+        # Calculate shared secret based on own private key and peer public key.
+        shared_key_list = self.lesc_private_key.exchange(ec.ECDH(), lesc_peer_public_key_obj2)
+        shared_key_list = list(shared_key_list)
+        shared_key_list = shared_key_list[::-1]
+        logger.debug("shared_key_list = {}".format(shared_key_list))
+
+        logger.debug("Shared secret list, little endian: {} \n".format(" ".join(["0x{:02X}".format(i) for i in shared_key_list])))
+        logger.debug("len(shared_key_list) = {}".format(len(shared_key_list)))
+        # Reply to Softdevice with shared key
+        lesc_dhkey_array = util.list_to_uint8_array(shared_key_list)
+        logger.debug("lesc_dhkey_array: {}".format(lesc_dhkey_array))
+        lesc_dhkey = driver.ble_gap_lesc_dhkey_t()
+        lesc_dhkey.key = lesc_dhkey_array.cast()
+        return lesc_dhkey
 
     @NordicSemiErrorCheck(expected=BLEGapSecStatus.success)
     def authenticate(
