+++ tests/test_passkey.py
@@ -44,9 +44,6 @@ import random
 import string
 import logging
 
-from cryptography.hazmat.backends import default_backend
-from cryptography.hazmat.primitives.asymmetric import ec
-
 from pc_ble_driver_py.observers import BLEDriverObserver, BLEAdapterObserver
 from driver_setup import Settings, setup_adapter
 
@@ -68,47 +65,6 @@ logger = logging.getLogger(__name__)
 passkeyQueue = Queue()
 authStatusQueue = Queue()
 
-def change_endianness(input_string):
-    output = ""
-    for i in range(len(input_string) - 1, 0, -2):
-        output += input_string[i - 1]
-        output += input_string[i]
-    return output
-
-class PublicKey(object):
-
-    def __init__(self, x, y, curve_type):
-        assert isinstance(x, (int)), 'Invalid argument type'
-        assert isinstance(y, (int)), 'Invalid argument type'
-        self.x = x
-        self.y = y
-        self.curve_type = curve_type
-
-    def to_c(self):
-
-        logger.debug("ECDH key: Before: x: {:X}".format(self.x))
-        logger.debug("ECDH key: Before: y: {:X}".format(self.y))
-
-        x_list = self._int_to_list(self.x)[::-1]
-        y_list = self._int_to_list(self.y)[::-1]
-
-        logger.debug("ECDH key: After: x: {}".format(" ".join(["0x{:02X}".format(i) for i in x_list])))
-        logger.debug("ECDH key: After: y: {}".format(" ".join(["0x{:02X}".format(i) for i in y_list])))
-
-        return util.list_to_uint8_array(x_list + y_list)
-
-    def _int_to_list(self, input_integer):
-        output_list = []
-        input_hex_string = "{:X}".format(input_integer)
-
-        # Add zeros to start key if they are stripped.
-        while len(input_hex_string) < 64:
-            input_hex_string = "0" + input_hex_string
-
-        for i in range(1, len(input_hex_string), 2):
-            output_list.append(int((input_hex_string[i-1] + input_hex_string[i]), 16))
-
-        return output_list
 
 class Central(BLEDriverObserver, BLEAdapterObserver):
     def __init__(self, adapter):
@@ -174,51 +130,14 @@ class Central(BLEDriverObserver, BLEAdapterObserver):
         # Generate own private and public lesc keys if lesc is enabled.
         if peer_params.lesc and self.lesc:
             logger.debug("Generate LE Secure Connection ECDH private and public key.")
-            self.lesc_private_key = ec.generate_private_key(ec.SECP256R1(), default_backend())
-            lesc_own_public_key = self.lesc_private_key.public_key()
-
-            # Put own lesc public key into keyset.
-            lesc_own_public_key_numbers = lesc_own_public_key.public_numbers()
-            lesc_own_public_key_array = PublicKey(lesc_own_public_key_numbers.x,
-                                                  lesc_own_public_key_numbers.y,
-                                                  lesc_own_public_key_numbers.curve).to_c()
-
-            self.adapter.lesc_pk_own.pk = lesc_own_public_key_array.cast()
-            lesc_own_public_key_list = util.uint8_array_to_list(self.adapter.lesc_pk_own.pk, 64)
-            logger.debug("lesc_own_public_key_list {}".format(lesc_own_public_key_list))
-            ble_driver.ble_gap_sec_params_reply(conn_handle, BLEGapSecStatus.success, None, None, lesc_own_public_key_list)
+            self.lesc_private_key = self.adapter.generate_lesc_private_key()
+            keyset = self.adapter.generate_lesc_keyset(self.lesc_private_key)
+            
+            ble_driver.ble_gap_sec_params_reply(conn_handle, BLEGapSecStatus.success, None, keyset)
 
     def on_gap_evt_lesc_dhkey_request(self, ble_driver, conn_handle, peer_public_key, oobd_req):
         logger.info("Central on_gap_evt_lesc_dhkey_request.")
-
-        self.lesc_pk_peer = peer_public_key
-        # Translate incoming peer public key to x an y components as integers.
-        peer_public_key_list = self.lesc_pk_peer.pk
-        peer_public_key_x = int(change_endianness("".join(["{:02X}".format(i) for i in peer_public_key_list[:32]])), 16)
-        peer_public_key_y = int(change_endianness("".join(["{:02X}".format(i) for i in peer_public_key_list[32:]])), 16)
-
-        logger.debug("Peer public DH key, big endian, x: 0x{:X}, y: 0x{:X}".format(peer_public_key_x, peer_public_key_y))
-
-        # Generate a _EllipticCurvePublicKey object of the received peer public key.
-        lesc_peer_public_key_obj = ec.EllipticCurvePublicNumbers(peer_public_key_x,
-                                                                 peer_public_key_y,
-                                                                 ec.SECP256R1())
-        lesc_peer_public_key_obj2 = lesc_peer_public_key_obj.public_key(default_backend())
-
-        # Calculate shared secret based on own private key and peer public key.
-        shared_key_list = self.lesc_private_key.exchange(ec.ECDH(), lesc_peer_public_key_obj2)
-        shared_key_list = list(shared_key_list)
-        shared_key_list = shared_key_list[::-1]
-        logger.debug("shared_key_list = {}".format(shared_key_list))
-
-        logger.debug("Shared secret list, little endian: {} \n".format(" ".join(["0x{:02X}".format(i) for i in shared_key_list])))
-        logger.debug("len(shared_key_list) = {}".format(len(shared_key_list)))
-        # Reply to Softdevice with shared key
-        lesc_dhkey_array = util.list_to_uint8_array(shared_key_list)
-        logger.debug("lesc_dhkey_array: {}".format(lesc_dhkey_array))
-        lesc_dhkey = driver.ble_gap_lesc_dhkey_t()
-        lesc_dhkey.key = lesc_dhkey_array.cast()
-
+        lesc_dhkey = self.adapter.generate_lesc_dhkey(peer_public_key)
         self.adapter.driver.ble_gap_lesc_dhkey_reply(conn_handle, lesc_dhkey)
 
     def on_gap_evt_auth_key_request(
@@ -258,7 +177,7 @@ class Peripheral(BLEDriverObserver, BLEAdapterObserver):
         self.conn_q.put(conn_handle)
 
     def on_gap_evt_sec_params_request(self, ble_driver, conn_handle, **kwargs):
-        logger.info("PERIPHERAL: on_gap_evt_sec_params_request.")
+        logger.info("Peripheral: on_gap_evt_sec_params_request.")
         sec_params = BLEGapSecParams.from_c(
             kwargs['peer_params']
         )
@@ -270,55 +189,19 @@ class Peripheral(BLEDriverObserver, BLEAdapterObserver):
         sec_params.kdist_peer = BLEGapSecKDist(True, True, False, False)
 
         logger.debug("Generate LE Secure Connection ECDH private and public key.")
-        self.lesc_private_key = ec.generate_private_key(ec.SECP256R1(), default_backend())
-        lesc_own_public_key = self.lesc_private_key.public_key()
-
-        # Put own lesc public key into keyset.
-        lesc_own_public_key_numbers = lesc_own_public_key.public_numbers()
-        lesc_own_public_key_array = PublicKey(lesc_own_public_key_numbers.x,
-                                              lesc_own_public_key_numbers.y,
-                                              lesc_own_public_key_numbers.curve).to_c()
-
-        self.adapter.lesc_pk_own.pk = lesc_own_public_key_array.cast()
-        lesc_own_public_key_list = util.uint8_array_to_list(self.adapter.lesc_pk_own.pk, 64)
-        logger.debug("lesc_own_public_key_list {}".format(lesc_own_public_key_list))
-
-        ble_driver.ble_gap_sec_params_reply(conn_handle, BLEGapSecStatus.success, sec_params, None, lesc_own_public_key_list)
+        self.adapter.generate_lesc_private_key()
+        keyset = self.adapter.generate_lesc_keyset()
+        
+        ble_driver.ble_gap_sec_params_reply(conn_handle, BLEGapSecStatus.success, sec_params, keyset)
 
     def on_gap_evt_lesc_dhkey_request(self, ble_driver, conn_handle, peer_public_key, oobd_req):
-        logger.info("PERIPHERAL: on_gap_evt_lesc_dhkey_request.")
+        logger.info("Peripheral: on_gap_evt_lesc_dhkey_request.")
+        lesc_dhkey = self.adapter.generate_lesc_dhkey(peer_public_key)
 
-        self.lesc_pk_peer = peer_public_key
-        # Translate incoming peer public key to x an y components as integers.
-        peer_public_key_list = self.lesc_pk_peer.pk
-        peer_public_key_x = int(change_endianness("".join(["{:02X}".format(i) for i in peer_public_key_list[:32]])), 16)
-        peer_public_key_y = int(change_endianness("".join(["{:02X}".format(i) for i in peer_public_key_list[32:]])), 16)
-
-        logger.debug("Peer public DH key, big endian, x: 0x{:X}, y: 0x{:X}".format(peer_public_key_x, peer_public_key_y))
-
-        # Generate a _EllipticCurvePublicKey object of the received peer public key.
-        lesc_peer_public_key_obj = ec.EllipticCurvePublicNumbers(peer_public_key_x,
-                                                                 peer_public_key_y,
-                                                                 ec.SECP256R1())
-        lesc_peer_public_key_obj2 = lesc_peer_public_key_obj.public_key(default_backend())
-
-        # Calculate shared secret based on own private key and peer public key.
-        shared_key_list = self.lesc_private_key.exchange(ec.ECDH(), lesc_peer_public_key_obj2)
-        shared_key_list = list(shared_key_list)
-        shared_key_list = shared_key_list[::-1]
-        logger.debug("shared_key_list = {}".format(shared_key_list))
-
-        logger.debug("Shared secret list, little endian: {} \n".format(" ".join(["0x{:02X}".format(i) for i in shared_key_list])))
-        logger.debug("len(shared_key_list) = {}".format(len(shared_key_list)))
-        # Reply to Softdevice with shared key
-        lesc_dhkey_array = util.list_to_uint8_array(shared_key_list)
-        logger.debug("lesc_dhkey_array: {}".format(lesc_dhkey_array))
-        lesc_dhkey = driver.ble_gap_lesc_dhkey_t()
-        lesc_dhkey.key = lesc_dhkey_array.cast()
         self.adapter.driver.ble_gap_lesc_dhkey_reply(conn_handle, lesc_dhkey)
 
     def on_gap_evt_passkey_display(self, ble_driver, conn_handle, passkey):
-        logger.info("PERIPHERAL on_gap_evt_passkey_display.")
+        logger.info("Peripheral on_gap_evt_passkey_display.")
         passkeyQueue.put(passkey)
 
     def on_gap_evt_auth_status(
@@ -333,7 +216,7 @@ class Peripheral(BLEDriverObserver, BLEAdapterObserver):
                         kdist_peer,
                         auth_status
                     ):
-        logger.info("PERIPHERAL on_gap_evt_auth_status.")
+        logger.info("Peripheral on_gap_evt_auth_status.")     
         authStatusQueue.put(auth_status)
